Move & untangle transform_tree
