+++ crates/store/re_tf/Cargo.toml
@@ -25,7 +25,7 @@ re_entity_db.workspace = true # It would be nice not to depend on this, but we n
 re_log_types.workspace = true
 re_log.workspace = true
 re_tracing.workspace = true
-re_types.workspace = true
+re_types = { workspace = true, features = ["glam"] }
 
 ahash.workspace = true
 bitflags.workspace = true
@@ -33,4 +33,4 @@ glam.workspace = true
 itertools.workspace = true
 nohash-hasher.workspace = true
 smallvec.workspace = true
-vec1.workspace = true
+vec1 = { workspace = true, features = ["smallvec-v1"] }
