+++ crates/viewer/re_view_spatial/src/contexts/mod.rs
@@ -2,9 +2,11 @@ mod depth_offsets;
 mod transform_tree_context;
 
 pub use depth_offsets::EntityDepthOffsets;
+use re_tf::TransformInfo;
 use re_types::ViewClassIdentifier;
 use re_view::AnnotationSceneContext;
-pub use transform_tree_context::{TransformInfo, TransformTreeContext, TwoDInThreeDTransformInfo};
+
+pub use transform_tree_context::TransformTreeContext;
 
 // -----------------------------------------------------------------------------
