+++ crates/viewer/re_view_spatial/src/space_camera_3d.rs
@@ -2,9 +2,10 @@ use glam::Vec3;
 use macaw::IsoTransform;
 
 use re_log_types::EntityPath;
+use re_tf::image_view_coordinates;
 use re_types::components::ViewCoordinates;
 
-use crate::{Pinhole, visualizers::image_view_coordinates};
+use crate::Pinhole;
 
 /// A logged camera that connects spaces.
 #[derive(Clone, PartialEq)]
